For bcdToLong, use unsigned long instead of UInt32 to make it compile on Windows.
git-svn-id: https://frontierkernel.svn.sourceforge.net/svnroot/frontierkernel/Frontier/trunk@237 4e289971-d10f-0410-980f-25908192c0f7

+++ Common/headers/ops.h
@@ -108,4 +108,4 @@ extern void getsystemversionstring (bigstring, bigstring);
 
 extern void getsizestring (unsigned long, bigstring);
 
-extern UInt32 bcdToLong (UInt32); /* 2004-11-16 creedon */
+extern unsigned long bcdToLong (unsigned long); /* 2004-11-16 creedon */
